Refactor backend for names
I refactored a ton of code out of main.py so that it would be able to
generate name labels as well as property tags (though not on the
same page). The code is now more organized and only routes are in main.py.

The Frontend does not have support yet.