Refactor Frontend for Names
I added support for name label generation to the frontend in a separate tab on the
same single page. I refactored the one page into a bunch of templates for
cleanliness (one for each form, submit section, and javascript section, and a nav page).
Basically the names are just a simpler version of numbers. Single names can
be added, or multiple names pasted in on separate lines.

I also disabled Spruce store persistence since I wasn't using it and it introduced
a bug that broke page loading.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>