Overflow fix for names, whitespace stripping, and sorting
Some names were overflowing with my arbitrarily chosen width. Should
be fixed now, will err on the side of wrapping too early.

Whitespace is now cleaned in javascript, replacing all tabs with spaces and
any multiple spaces with single spaces. Not as clean as it would be in python
but it makes them display correctly on the frontend for better UX.

Names and numbers are now sorted (python built-in) before creating labels so
that the order is a bit less totally random

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>